requirements.txt
django==5.1.5
pyttsx3==2.90
SpeechRecognition==3.10.1
google-generativeai==0.3.2
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9 